+++ tests/FSharpLint.Console.Tests/Attributes.fs
@@ -1,0 +1,7 @@
+namespace FSharpLint.Console.Tests
+
+open NUnit.Framework
+
+[<assembly: Parallelizable(ParallelScope.Fixtures)>]
+
+do ()